godoc

+++ backend/Classes/Commands/Mkfs.go
@@ -1,3 +1,5 @@
+// Package Commands provides implementations for various filesystem commands,
+// including commands for formatting partitions and creating EXT2 filesystems.
 package Commands
 
 import (
@@ -13,53 +15,64 @@ import (
 	"time"
 )
 
+// Mkfs represents the MKFS command, which is used to format a mounted partition
+// with an EXT2 filesystem. It embeds the common CommandStruct for shared metadata
+// and includes a map for command parameters.
 type Mkfs struct {
-	Interfaces.CommandStruct
-	Params map[string]string
+	Interfaces.CommandStruct                   // Embedded command structure (provides Type, Line, Column, etc.)
+	Params                   map[string]string // Params contains command parameters such as "id" and "type".
 }
 
+// NewMkfs creates a new Mkfs command instance with the specified source location and parameters.
 func NewMkfs(line, column int, params map[string]string) *Mkfs {
 	return &Mkfs{
 		CommandStruct: Interfaces.CommandStruct{
-			Type:   Utils.MKFS,
-			Line:   line,
-			Column: column,
+			Type:   Utils.MKFS, // Command type constant for MKFS.
+			Line:   line,       // Source code line where the command is defined.
+			Column: column,     // Source code column where the command is defined.
 		},
 		Params: params,
 	}
 }
 
+// Exec executes the MKFS command to format the partition with an EXT2 filesystem.
+// It validates parameters, verifies the mounted partition, calculates filesystem parameters,
+// creates a new superblock, initializes inodes and blocks, creates root and users files,
+// updates bitmaps, and writes the final superblock back to disk.
 func (m *Mkfs) Exec() {
+	// Validate command parameters.
 	if err := m.validateParams(); err != nil {
 		m.AppendError(err.Error())
 		return
 	}
 
+	// Start building a console log for the operation.
 	var consoleString string
 	consoleString = "=================MKFS=================\n"
 	consoleString += fmt.Sprintf("Id: %s, Type: %s, Fs: EXT2\n", m.Params["id"], m.Params["type"])
 
-	// Check if the given id belong to a partition that is mounted
+	// Retrieve the mounted partition using the provided partition id.
 	var mountedPartition *Structs.MountedPartition
 	if mountedPartition = m.getPartition(); mountedPartition == nil {
 		m.AppendError("Partition ID not found")
 		return
 	}
 
-	// Check if mounted partition has the status of mounted
+	// Ensure that the partition is mounted.
 	if mountedPartition.Status != [1]byte{'1'} {
 		m.AppendError("Partition is not mounted")
 		return
 	}
 
-	// Checking if the mounted partition is the same as the partition in disc
+	// Verify that the mounted partition exists in the disk's MBR.
 	mbrPartition := m.checkMountedPartitionHealth(mountedPartition, consoleString)
 	if mbrPartition == nil {
 		m.AppendError("Mounted partition does not exist at disc MBR")
 		return
 	}
 
-	// Calculating the size of the inodes
+	// Calculate the number of inodes (n) based on the partition size minus the superblock size.
+	// The denominator is calculated based on overhead (4 bytes) plus sizes of an inode and three file blocks.
 	numerator := mbrPartition.Size - int32(binary.Size(Structs.SuperBlock{}))
 	baseDenominator := 4 + int32(binary.Size(Structs.Inode{})) + 3*int32(binary.Size(Structs.FileBlock{}))
 	var temp int32 = 0
@@ -67,19 +80,24 @@ func (m *Mkfs) Exec() {
 	n := int32(numerator / denominator)
 	consoleString += fmt.Sprintf("I-NODES qnt: %d\n", n)
 
+	// Create the EXT2 filesystem structure.
 	resultString := m.makeEXT2(n, mbrPartition, mountedPartition, consoleString)
 	if resultString == nil {
 		return
 	}
 
 	consoleString = *resultString
-	// Only FS2 filesystem right now
+	// Append footer to the console output.
 	consoleString += "=================END MKFS=================\n"
 	m.LogConsole(consoleString)
 }
 
+// makeEXT2 creates an EXT2 filesystem on the given partition.
+// It builds a new superblock, calculates starting offsets for the bitmaps,
+// inode area, and block area, initializes inodes and file blocks, and creates
+// the root folder and users file. It returns a pointer to the updated console log string.
 func (m *Mkfs) makeEXT2(n int32, mbrPartition *Structs.Partition, mountedPartition *Structs.MountedPartition, consoleString string) *string {
-	// Creating super block
+	// Initialize a new superblock with EXT2 settings.
 	var superBlock Structs.SuperBlock
 	superBlock.FilesystemType = 2
 	superBlock.InodesCount = n
@@ -93,7 +111,7 @@ func (m *Mkfs) makeEXT2(n int32, mbrPartition *Structs.Partition, mountedPartiti
 	superBlock.InodeSize = int32(binary.Size(Structs.Inode{}))
 	superBlock.BlockSize = int32(binary.Size(Structs.FileBlock{}))
 
-	// Calculate start position
+	// Calculate starting offsets for bitmaps, inodes, and file blocks.
 	superBlock.BmInodeStart = mbrPartition.Start + int32(binary.Size(Structs.SuperBlock{}))
 	superBlock.BmBlockStart = superBlock.BmInodeStart + n
 	superBlock.InodeStart = superBlock.BmBlockStart + 3*n
@@ -102,6 +120,7 @@ func (m *Mkfs) makeEXT2(n int32, mbrPartition *Structs.Partition, mountedPartiti
 	consoleString += "_________________Creating EXT2_________________\n"
 	consoleString += superBlock.ToString() + "\n\n"
 
+	// Open the disk file for writing.
 	file, err := Utils.OpenFile(mountedPartition.Path)
 	if err != nil {
 		m.AppendError(err.Error())
@@ -109,6 +128,7 @@ func (m *Mkfs) makeEXT2(n int32, mbrPartition *Structs.Partition, mountedPartiti
 	}
 	defer file.Close()
 
+	// Initialize the block bitmap: write 0 for each block (total of 3*n blocks).
 	for i := int32(0); i < 3*n; i++ {
 		if err := Utils.WriteObject(file, byte(0), int64(superBlock.BmInodeStart)); err != nil {
 			m.AppendError(err.Error())
@@ -116,34 +136,34 @@ func (m *Mkfs) makeEXT2(n int32, mbrPartition *Structs.Partition, mountedPartiti
 		}
 	}
 
-	// initialize inodes and blocks to file
+	// Initialize all inodes and file blocks to default values.
 	if err := initInodesAndBlocks(n, superBlock, file); err != nil {
 		m.AppendError(err.Error())
 		return nil
 	}
 
-	// Create the root folder and the users file
+	// Create the root folder and the users file in the filesystem.
 	if err := createRootAndUsersFile(superBlock, file); err != nil {
 		m.AppendError(err.Error())
 		return nil
 	}
+	// Update free counts (subtract reserved entries for root and users).
 	superBlock.FreeBlocksCount = 3*n - 2
 	superBlock.FreeInodesCount = n - 2
 
-	// Write super block at the start of the partition
+	// Write the updated superblock at the start of the partition.
 	if err := Utils.WriteObject(file, superBlock, int64(mbrPartition.Start)); err != nil {
 		m.AppendError(err.Error())
 		return nil
 	}
 
-	// Mark the first inodes and blocks as used
+	// Mark reserved inodes and blocks as used.
 	if err := markUsedInodesAndNodes(file, superBlock); err != nil {
 		m.AppendError(err.Error())
 		return nil
 	}
 
-	// printing the list of inodes in partition
-	// Just printing the first 3 since the first 2 are the users and root folder and the third one is the same for the rest
+	// Print the first few inodes for debugging purposes.
 	consoleString += "\tPrinting Inodes:\n"
 	for i := int32(0); i < 3; i++ {
 		var inode Structs.Inode
@@ -157,7 +177,7 @@ func (m *Mkfs) makeEXT2(n int32, mbrPartition *Structs.Partition, mountedPartiti
 	}
 	consoleString += "\t\t......\n\n"
 
-	// Printing the only folder block available at the moment
+	// Print the first folder block.
 	consoleString += "\tPrinting Folder Block 0:\n"
 	var folderBlock Structs.FolderBlock
 	if err := Utils.ReadObject(file, &folderBlock, int64(superBlock.BlockStart)); err != nil {
@@ -166,7 +186,7 @@ func (m *Mkfs) makeEXT2(n int32, mbrPartition *Structs.Partition, mountedPartiti
 	}
 	consoleString += folderBlock.ToString() + "\n"
 
-	// printing the only file block at the moment, next to the first folder block
+	// Print the first file block (located after the folder block).
 	consoleString += "\n\tPrinting File Block 0:\n"
 	var fileBlock Structs.FileBlock
 	offset := int64(superBlock.BlockStart + int32(binary.Size(Structs.FolderBlock{})))
@@ -182,71 +202,76 @@ func (m *Mkfs) makeEXT2(n int32, mbrPartition *Structs.Partition, mountedPartiti
 	return &consoleString
 }
 
+// markUsedInodesAndNodes marks reserved inodes and blocks as used in the filesystem's bitmaps.
+// It writes a byte value of 1 for the first two inodes and the first two blocks.
 func markUsedInodesAndNodes(file *os.File, superBlock Structs.SuperBlock) error {
-	// Marking the first inode (root inode) as used in the first item of the bitmap
+	// Mark the first inode (root inode) as used.
 	if err := Utils.WriteObject(file, byte(1), int64(superBlock.BmInodeStart)); err != nil {
 		return err
 	}
-	// Marking the second inode (users.txt inode) as used in the second item of the bitmap
+	// Mark the second inode (users.txt inode) as used.
 	if err := Utils.WriteObject(file, byte(1), int64(superBlock.BmInodeStart+1)); err != nil {
 		return err
 	}
-	// Marking the first block (root folder block) as used in the first item of the bitmap
+	// Mark the first block (root folder block) as used.
 	if err := Utils.WriteObject(file, byte(1), int64(superBlock.BmBlockStart)); err != nil {
 		return err
 	}
-	// Marking the second block (users.txt file block) as used in the second item of the bitmap
+	// Mark the second block (users.txt file block) as used.
 	if err := Utils.WriteObject(file, byte(1), int64(superBlock.BmBlockStart+1)); err != nil {
 		return err
 	}
 	return nil
 }
 
+// createRootAndUsersFile creates the root directory and the users file ("/users.txt").
+// It creates corresponding inodes, a folder block for the root directory, and a file block for users data.
 func createRootAndUsersFile(superBlock Structs.SuperBlock, file *os.File) error {
+	// Create the root inode (type folder) and the users inode (type file).
 	rootInode := Structs.NewInode([1]byte{0})
 	usersInode := Structs.NewInode([1]byte{1})
 
-	// this is set to 0 because we are using the first folder Block available
+	// Set the first block pointers:
+	// Root inode points to folder block 0, and users inode points to file block 1.
 	rootInode.IBlock[0] = 0
-	// This is set to 1 because here we are pointing to the next file block
 	usersInode.IBlock[0] = 1
 
-	// Assign the size of the content of the file into the users Inode
+	// Initialize the users file content.
 	data := "1,G,root\n1,U,root,root,123\n"
 	actualSize := int32(len(data))
 	usersInode.Size = actualSize
 
-	// File block for storing the data of the users
+	// Create a file block for the users file and copy the data.
 	var usersFileBlock Structs.FileBlock
 	copy(usersFileBlock.Content[:], data[:])
 
-	// Creating the root folder block for "/"
+	// Create a folder block for the root directory ("/").
 	var rootFolderBlock Structs.FolderBlock
-	// The current folder is "." referring to "/"
+	// Entry for current directory "."
 	rootFolderBlock.Content[0].Inode = 0
 	copy(rootFolderBlock.Content[0].Name[:], ".")
-	// The parent folder is "." too, so still referring to "/"
+	// Entry for parent directory ".." (same as root in this case)
 	rootFolderBlock.Content[1].Inode = 0
 	copy(rootFolderBlock.Content[1].Name[:], "..")
-	// The next available space is for the file users.txt
+	// Entry for the users file "users.txt"
 	rootFolderBlock.Content[2].Inode = 1
 	copy(rootFolderBlock.Content[2].Name[:], "users.txt")
-	// update the last inode pointer
+	// Mark remaining entry as unused.
 	rootFolderBlock.Content[3].Inode = -1
 
-	// Writing the root inode at the start of the Inode part
+	// Write the root inode at the beginning of the inode area.
 	if err := Utils.WriteObject(file, rootInode, int64(superBlock.InodeStart)); err != nil {
 		return err
 	}
-	// Writing the users inode next to the first inode
+	// Write the users inode right after the root inode.
 	if err := Utils.WriteObject(file, usersInode, int64(superBlock.InodeStart+int32(binary.Size(Structs.Inode{})))); err != nil {
 		return err
 	}
-	// Writing the root folder block
+	// Write the root folder block at the beginning of the block area.
 	if err := Utils.WriteObject(file, rootFolderBlock, int64(superBlock.BlockStart)); err != nil {
 		return err
 	}
-	// Writing the users file block next to the root folder block
+	// Write the users file block next to the folder block.
 	if err := Utils.WriteObject(file, usersFileBlock, int64(superBlock.BlockStart+int32(binary.Size(Structs.FolderBlock{})))); err != nil {
 		return err
 	}
@@ -254,19 +279,26 @@ func createRootAndUsersFile(superBlock Structs.SuperBlock, file *os.File) error 
 	return nil
 }
 
+// initInodesAndBlocks initializes all inodes and file blocks on the filesystem.
+// It writes a default inode (with all IBlock pointers set to -1) for each inode slot,
+// and writes a default (empty) file block for each of the 3*n file blocks.
 func initInodesAndBlocks(n int32, superBlock Structs.SuperBlock, file *os.File) error {
+	// Prepare a default inode with IBlock pointers initialized to -1.
 	var newInode Structs.Inode
-	for i := int32(0); i < 15; i++ {
+	for i := 0; i < 15; i++ {
 		newInode.IBlock[i] = -1
 	}
 
+	// Write n inodes to the inode area.
 	for i := int32(0); i < n; i++ {
 		if err := Utils.WriteObject(file, newInode, int64(superBlock.InodeStart+i*int32(binary.Size(Structs.Inode{})))); err != nil {
 			return err
 		}
 	}
 
+	// Prepare a default file block.
 	var newFileBlock Structs.FileBlock
+	// Write 3*n file blocks to the block area.
 	for i := int32(0); i < 3*n; i++ {
 		if err := Utils.WriteObject(file, newFileBlock, int64(superBlock.BlockStart+i*int32(binary.Size(Structs.FileBlock{})))); err != nil {
 			return err
@@ -275,6 +307,9 @@ func initInodesAndBlocks(n int32, superBlock Structs.SuperBlock, file *os.File) 
 	return nil
 }
 
+// checkMountedPartitionHealth verifies that the mounted partition exists on disk by reading the MBR.
+// It opens the disk file for the mounted partition, reads the MBR, logs the MBR info to the console string,
+// and searches for the partition that matches the mounted partition's ID.
 func (m *Mkfs) checkMountedPartitionHealth(mountedPartition *Structs.MountedPartition, consoleString string) *Structs.Partition {
 	file, err := Utils.OpenFile(mountedPartition.Path)
 	if err != nil {
@@ -288,8 +323,8 @@ func (m *Mkfs) checkMountedPartitionHealth(mountedPartition *Structs.MountedPart
 		return nil
 	}
 	consoleString += "Founded MBR:\n" + discMBR.ToString() + "\n"
+	// Search through the MBR partitions for the one that matches the mounted partition's ID.
 	for _, partition := range discMBR.Partitions {
-		//fmt.Println(partition.ToString())
 		if partition.Id == mountedPartition.Id {
 			return &partition
 		}
@@ -298,6 +333,9 @@ func (m *Mkfs) checkMountedPartitionHealth(mountedPartition *Structs.MountedPart
 	return nil
 }
 
+// validateParams ensures that the required parameters for the MKFS command are provided.
+// It checks that the "id" parameter exists and sets a default "type" if not provided.
+// It also converts the "id" parameter to uppercase.
 func (m *Mkfs) validateParams() error {
 	if _, ok := m.Params["id"]; !ok {
 		return errors.New("missing parameter -id")
@@ -309,6 +347,8 @@ func (m *Mkfs) validateParams() error {
 	return nil
 }
 
+// getPartition searches for a mounted partition that matches the provided "id" parameter.
+// It iterates over the global list of mounted partitions and returns the matching partition.
 func (m *Mkfs) getPartition() *Structs.MountedPartition {
 	for _, part := range env.GetPartitions() {
 		if string(part.Id[:]) == m.Params["id"] {
